Remove AnalyticExt
With the AnalyticExts, precompilation fails. I could not fix this in any
other way than moving the code for extensions into the other extensions
