+++ docs/make.jl
@@ -47,11 +47,9 @@ makedocs(
         Base.get_extension(ClimaUtilities, :DataHandlingExt),
         Base.get_extension(ClimaUtilities, :InterpolationsRegridderExt),
         Base.get_extension(ClimaUtilities, :NCFileReaderExt),
-        Base.get_extension(ClimaUtilities, :SpaceVaryingInputsAnalyticExt),
         Base.get_extension(ClimaUtilities, :SpaceVaryingInputsExt),
         Base.get_extension(ClimaUtilities, :TempestRegridderExt),
         Base.get_extension(ClimaUtilities, :TimeVaryingInputs0DExt),
-        Base.get_extension(ClimaUtilities, :TimeVaryingInputsAnalyticExt),
         Base.get_extension(ClimaUtilities, :TimeVaryingInputsExt),
     ],
 )
