--- ext/TimeVaryingInputsAnalyticExt.jl
@@ -1,36 +1,0 @@
-module TimeVaryingInputsAnalyticExt
-
-import ClimaUtilities.TimeVaryingInputs
-
-
-import ClimaCore
-
-struct AnalyticTimeVaryingInput{F <: Function} <:
-       TimeVaryingInputs.AbstractTimeVaryingInput
-    # func here has to be GPU-compatible (e.g., splines are not) and reasonably fast (e.g.,
-    # no large allocations)
-    func::F
-end
-
-# _kwargs... is needed to seamlessly support the other TimeVaryingInputs.
-function TimeVaryingInputs.TimeVaryingInput(
-    input::Function;
-    method = nothing,
-    _kwargs...,
-)
-    isnothing(method) ||
-        @warn "Interpolation method is ignored for analytical functions"
-    return AnalyticTimeVaryingInput(input)
-end
-
-function TimeVaryingInputs.evaluate!(
-    dest,
-    input::AnalyticTimeVaryingInput,
-    time,
-    args...;
-    kwargs...,
-)
-    dest .= input.func(time, args...; kwargs...)
-    return nothing
-end
-end
