+++ ext/TimeVaryingInputsExt.jl
@@ -17,6 +17,38 @@ import ClimaUtilities.DataHandling
 import ClimaUtilities.DataHandling:
     DataHandler, previous_time, next_time, regridded_snapshot, available_times
 
+# Ideally, we should be able to split off the analytic part in a different
+# extension, but precompilation stops working when we do so
+
+struct AnalyticTimeVaryingInput{F <: Function} <:
+       TimeVaryingInputs.AbstractTimeVaryingInput
+    # func here has to be GPU-compatible (e.g., splines are not) and reasonably fast (e.g.,
+    # no large allocations)
+    func::F
+end
+
+# _kwargs... is needed to seamlessly support the other TimeVaryingInputs.
+function TimeVaryingInputs.TimeVaryingInput(
+    input::Function;
+    method = nothing,
+    _kwargs...,
+)
+    isnothing(method) ||
+        @warn "Interpolation method is ignored for analytical functions"
+    return AnalyticTimeVaryingInput(input)
+end
+
+function TimeVaryingInputs.evaluate!(
+    dest,
+    input::AnalyticTimeVaryingInput,
+    time,
+    args...;
+    kwargs...,
+)
+    dest .= input.func(time, args...; kwargs...)
+    return nothing
+end
+
 """
     InterpolatingTimeVaryingInput23D
